+++ backend/backend/settings.py
@@ -57,6 +57,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'users',
+    'tasks',
     'api',
     'rest_framework',
     'corsheaders',
