+++ backend/tasks/models.py
@@ -1,0 +1,13 @@
+from django.db import models
+from django.contrib.auth.models import User
+
+# Create your models here.
+class Task(models.Model):
+    title = models.CharField(max_length=200)
+    description = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    assigned_to = models.ForeignKey(User, on_delete=models.SET_NULL, null=True, related_name='+') # related_name='assigned_tasks')
+    author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='+') #  related_name='authored_tasks')
+
+    def __str__(self):
+        return self.title
