+++ backend/users/models.py
@@ -1,0 +1,10 @@
+from django.db import models
+from django.contrib.auth.models import User
+
+class Profile(models.Model):
+    user = models.OneToOneField(User, on_delete=models.CASCADE)
+    accept_tasks = models.BooleanField(default=False)
+    assign_tasks = models.BooleanField(default=False)
+
+    def __str__(self):
+        return f'{self.user.username} profile'
